# apps/api/openapi.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    MealNutrientInput:
      additionalProperties: false
      properties:
        amount:
          anyOf:
          - minimum: 0.0
            type: number
          - pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,10}|(?=[\d.]{1,15}0*$)\d{0,10}\.\d{0,4}0*$)
            type: string
          title: Amount
        code:
          maxLength: 40
          minLength: 1
          title: Code
          type: string
        confidence:
          anyOf:
          - maximum: 1.0
            minimum: 0.0
            type: number
          - pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,1}|(?=[\d.]{1,6}0*$)\d{0,1}\.\d{0,4}0*$)
            type: string
          - type: 'null'
          title: Confidence
        provenance:
          anyOf:
          - $ref: '#/components/schemas/NutrientProvenance'
          - type: 'null'
      required:
      - code
      - amount
      title: MealNutrientInput
      type: object
    MealSource:
      enum:
      - MANUAL
      - IMAGE
      - PDF
      - CHAT
      title: MealSource
      type: string
    MealType:
      enum:
      - BREAKFAST
      - LUNCH
      - DINNER
      - SNACKS
      title: MealType
      type: string
    MealUpsertRequest:
      additionalProperties: false
      properties:
        foodName:
          maxLength: 200
          minLength: 1
          title: Foodname
          type: string
        mealType:
          $ref: '#/components/schemas/MealType'
        notes:
          anyOf:
          - maxLength: 1000
            type: string
          - type: 'null'
          title: Notes
        nutrients:
          items:
            $ref: '#/components/schemas/MealNutrientInput'
          maxItems: 25
          minItems: 1
          title: Nutrients
          type: array
        occurredAt:
          format: date-time
          title: Occurredat
          type: string
        quantity:
          $ref: '#/components/schemas/QuantityInput'
        source:
          $ref: '#/components/schemas/MealSource'
          default: MANUAL
        sourceExtractionId:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Sourceextractionid
        timezone:
          title: Timezone
          type: string
      required:
      - mealType
      - foodName
      - quantity
      - occurredAt
      - timezone
      - nutrients
      title: MealUpsertRequest
      type: object
...
    NutrientProvenance:
      enum:
      - USER
      - LABEL_AI
      - PLATE_AI
      - PDF_AI
      title: NutrientProvenance
      type: string
...
    QuantityInput:
      additionalProperties: false
      properties:
        description:
          anyOf:
          - maxLength: 200
            type: string
          - type: 'null'
          title: Description
        unit:
          maxLength: 30
          minLength: 1
          title: Unit
          type: string
        value:
          anyOf:
          - exclusiveMinimum: 0.0
            type: number
          - pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,8}|(?=[\d.]{1,13}0*$)\d{0,8}\.\d{0,4}0*$)
            type: string
          title: Value
      required:
      - value
      - unit
      title: QuantityInput
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  title: Personal Calorie Tracker API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/meal-entries:
    get:
      operationId: listMealEntries
      parameters:
      - in: query
        name: dateFrom
        required: false
        schema:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Datefrom
      - in: query
        name: dateTo
        required: false
        schema:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Dateto
      - in: query
        name: page
        required: false
        schema:
          default: 1
          title: Page
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 20
          title: Limit
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - items:
                  type: string
                type: array
              - type: 'null'
              title: Mealtype
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Meals
      tags:
      - meals
    post:
      operationId: createMealEntry
      parameters:
      - in: header
        name: Idempotency-Key
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency-Key
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MealUpsertRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Meal
      tags:
      - meals
  /api/v1/meal-entries/{meal_id}:
    delete:
      operationId: deleteMealEntry
      parameters:
      - in: path
        name: meal_id
        required: true
        schema:
          format: uuid
          title: Meal Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Meal
      tags:
      - meals
    get:
      operationId: getMealEntry
      parameters:
      - in: path
        name: meal_id
        required: true
        schema:
          format: uuid
          title: Meal Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Meal
      tags:
      - meals
    put:
      operationId: replaceMealEntry
      parameters:
      - in: path
        name: meal_id
        required: true
        schema:
          format: uuid
          title: Meal Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MealUpsertRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Replace Meal
      tags:
      - meals
...